Discover view modes (#331)
* Move query params config to its own file

* Add a component to switch between display modes

* Make Hit component render configurable

* Tweak styling of the compact hit

* Add list view to discover

* Persist display mode on query string

* Share common components among hits components

* Improve highlighting of compact view results

* Use container fluid properly

* Set fluid container to use xl breakpoint
