+++ site/gatsby-site/src/components/discover/Filters.js
@@ -5,6 +5,7 @@ import Filter from './Filter';
 import Stats from './Stats';
 import styled from 'styled-components';
 import ClearFilters from './ClearFilters';
+import DisplayModeSwitch from './DisplayModeSwitch';
 
 const StyledFilter = styled(Filter)`
   .dropdown-toggle {
@@ -25,6 +26,7 @@ function Filters() {
       <Row className="d-none d-md-flex">
         <Col className="d-flex align-items-center">
           <Stats />
+          <DisplayModeSwitch />
         </Col>
         <Col className="d-flex justify-content-end">
           <ClearFilters>Clear Filters</ClearFilters>
