+++ site/gatsby-site/src/components/discover/hits/index.js
@@ -1,0 +1,5 @@
+import * as details from './Details';
+import * as compact from './Compact';
+import * as list from './List';
+
+export default { details, compact, list };
